Motor direction
All motors are correct direction. Note: For some reason the back right motor doesn't reverse direction in motor definition at line 164 (ie adding true doesn't reverse it). Idk why, must be a hardware issue or something. The only way is to change the sign of the speed (ie -90 instead of 90 to reverse direction)